Moved CursorChangeEvent to ObjectClicker
