--- src/l3d/src/events/CursorChangeEvent.ts
@@ -1,13 +1,0 @@
-module l3d {
-
-    export interface CursorChangeEvent extends Event {
-
-        cursor ?: string;
-
-    }
-
-    export class Dummy {}
-
-}
-
-export = l3d;
