+++ src/l3d/src/l3d.ts
@@ -18,5 +18,4 @@ export { FixedCamera } from './cameras/FixedCamera';
 export { BaseCamera } from './cameras/BaseCamera';
 export { Camera } from './cameras/Camera';
 export { ReplayCamera } from './cameras/ReplayCamera';
-export { CursorChangeEvent } from './events/CursorChangeEvent';
-export { Dummy } from './events/CursorChangeEvent';
+export { CursorChangeEvent } from './utils/ObjectClicker';
